requirements.txt
# client
requests==2.31.0
wbdata==0.3.0
google-cloud-translate==3.11.1
pycountry==22.3.5

# compute
numpy==1.25.0
xarray==2023.5.0
zarr==2.15.0
dask==2023.6.0
scipy==1.10.1
matplotlib==3.7.1

# Web
fastapi==0.97.0
uvicorn==0.22.0
gunicorn==20.1.0
PyJWT==2.8.0
jwcrypto==1.5.0

# AWS Infra
boto3==1.26.158
botocore==1.29.158
psycopg==3.1.9
